Remove merge conflict markers from README
# Description
There were an obvious merge conflict markers in the README.md (sections starting with `<<<<<< HEAD`) and a seemingly outdated changelog. To keep the README clean, I propose to remove this section.

However, the section was supposed to be dedicated to sponsors and supporters and I could not find an up-to-date list of sponsors to include there. I believe this is up to @MrNeRF to figure out.

Since there were no sponsors, I removed the whole section, but we should obviously bring it back if there are any.